refactor(workspace): P5.0+P5.1 — contracts/ tier scaffold + social-graph-api pilot
Begin the contracts tier: a single proto IDL root + per-service generated stub
crates, replacing each service's local build.rs proto compilation.

P5.0 — scaffolding:
- crates/contracts/proto/ becomes the single source of truth for IDL; moved
  social_graph/v1/*.proto here via `git mv` (history preserved).
- Added crates/contracts/proto/buf.yaml (lint now; breaking-change CI later).

P5.1 — social-graph-api pilot (proves the whole pattern end-to-end):
- New crate crates/contracts/social-graph-api: build.rs compiles social_graph.v1
  from ../proto with BOTH server + client stubs + the reflection descriptor;
  lib.rs re-exports the generated module and FILE_DESCRIPTOR_SET.
- social-graph (server) now deps social-graph-api: dropped its build.rs +
  build-deps, aliased `pub use social_graph_api as proto` (so every `proto::…`
  site is unchanged), and FILE_DESCRIPTOR_SET re-exports the api const.
- timeline (client) now deps social-graph-api: removed the cross-service proto
  recompile from its build.rs and the local `include_proto!`, aliased
  `use social_graph_api as sg_proto`. The fragile `../social-graph/proto/`
  reach-across is gone — the cross-service duplication this tier targets.

Pilot chosen to exercise server + client in one shot. The other 8 services keep
their local proto/build.rs untouched (P5.2). Green: `cargo check --workspace` and
`cargo test --workspace --no-run`; clippy clean on the new/touched crates.
