+++ crates/contracts/proto/social_graph/v1/messages.proto
