+++ crates/contracts/proto/social_graph/v1/service.proto
