+++ crates/contracts/social-graph-api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name        = "social-graph-api"
+version.workspace    = true
+edition.workspace    = true
+license.workspace    = true
+authors.workspace    = true
+repository.workspace = true
+description = "Generated gRPC contract for social_graph.v1 — server + client stubs and reflection descriptor."
+
+[dependencies]
+tonic       = { workspace = true }
+tonic-prost = { workspace = true }
+prost       = { workspace = true }
+prost-types = { workspace = true }
+
+[build-dependencies]
+tonic-prost-build = { workspace = true }
